scrapy-email-service/requirements.txt
Flask==3.0.0
Scrapy==2.11.0
dnspython==2.4.2
requests==2.31.0
twisted==23.10.0
